refactor inspector controlls into different files
